Collection of changes and bug fixes requested over the past couple of weeks.
TypeConverter.java: Handle Vectors as a Notes data type. The
specific case encountered is that multiple-valued columns in a
ViewNavigator are expanded to one ViewEntry per value when
iterating but with the column value returned as a Vector with one
element rather than a String.

NotesUserGroupManager.java: If unable to look up a group name,
log a warning and return rather than continuing. The specific
case encountered is a customer with a "/" character in a group
name.

NotesCrawlerThread.java, NotesConnectorDocument.java: Store the
filename of an attachment as a new document metadata property in
the attachment document sent to the GSA.

NotesAuthenticationManager.java: Fixed nesting error in group
prefixing code. Modified prefixed group list to send lower-case
group names since we send them as lower-case in the Policy ACL
and the GSA is case-sensitive when comparing group names for
authorization.

NotesDatabasePoller.java: When resetting the databases, reset the
ACL as well to force that to be updated.

http://codereview.appspot.com/5576058/