Inheritable ACL support.
The connector uses the TraversalContext.supportsInheritedAcls
method to determine whether the GSA supports inheritable
ACLs. When a database is configured to use ACL authorization, the
connector will send a policy ACL for the database if inheritable
ACLs are not supported. If inheritable ACLs are supported, the
connector will create an ACL for the database and add the
appropriate ACL fields to each document.

Database ACLs: NotesDatabasePoller will create two crawl
documents for each database processed when the database ACL has
been updated. These documents will contain the ACL information
necessary for the database (permit users, permit groups, deny
users, deny groups). There are two ACLs for each database in
order to support two inheritance models for documents. One
database ACL will have its inheritance type set to
PARENTOVERRIDES. This ACL will be used to control access to
documents with no readers fields for which the database access
permissions are the entire access control mechanism. The other
database ACL will have its inheritance type set to ANDBOTH. This
ACL will be used to control access to documents with readers
fields for which both the database access permissions and the
document readers must be consulted to determine document
access. NotesConnectorDocument now has support for constructing
database ACL documents.

Document ACLs: NotesConnectorDocument will add ACL fields to the
documents as they're being prepared to be sent to the GSA. When a
document has readers, the readers will be added as permitted
users/groups.

Roles: The connector will add a user's assigned roles to the user
cache as extra elements in the user's groups list. Each role will
be prefixed with the replica id of the database in which the role
is assigned, for example, "1234567890123456/[readerrole]". The
connector will cache roles assigned to groups in a separate field
in the group record using the same format. When constructing
document ACL fields, roles in the Readers list will be included
in the PERMIT groups for the document.

When doing group resolution, the connector begins with the user's
explicitly assigned groups and roles, and then adds any roles
assigned to the user's groups. For example, if a user is a member
of the "Employee" group, and the "Employee" group has the
"[worker]" role in database "12345", then the user's group list
will include "12345/[worker]".


NotesUserGroupManager.java: Added methods for dealing with
constructing user and group list for the GSA. Added methods to
update user and group records with role data and to retrieve role
data.

NotesConnectorDocument.java: Use helper classes
com.google.enterprise.connector.spi.SimpleDocument and
com.google.enterprise.connector.spi.SecureDocument to construct
documents for traversal. Added support for including database ACL
documents in the feed.

NotesDatabasePoller.java: Create database ACL documents for each
database crawled.

NotesConnectorDocumentList.java: Handle database ACL documents in
the document list.

http://codereview.appspot.com/5971057/