Changed user/group data from strings to Principal objects.
Users can be treated as either local (to Notes) or global
(authenticated by the GSA using an external system). A new
configuration option has been added to the connector
configuration form to indicate whether usernames used in the
connector should be treated as global. By default, in a
new connector installation users are considered global.

When users are global, the connector administrator is responsible
for updating the formula used to generate the GSA names cached in
the connector to match the names that will be used for login.

Groups are always treated as local.

NotesConnectorDocument.java no longer uses the SPI helpers
SimpleDocument/SecureDocument. SecureDocument doesn't support
Principal objects in the addPrincipal method, which was it's main
use here, so for now we're not using it. If SecureDocument is
enhanced in the future, we should reconsider this.



http://codereview.appspot.com/6195068/